GUI - V0.2
Added two new sub-procedures that enable the inputs of the user to be registered (getMouseMove) and then potential movements of that piece to be displayed by move indicators (displayMoves). I had struggled to create a terminating condition for the while loop that was displaying the potential moves, this was only to test the system worked but it frequently lead to crashes due to the fact I wasn't calling another mouse click and instead loading the previous location, meaning you could never click the close button to shut down the software. Thus once I added this feature I could test both sub-programmes and ensure they both functioned accordingly.