GUI - V0.21
Updated the move function such that the player can now click on each move indicator to move the piece, then I created a function where both the spriteboard and the main board update, moving the piece on the GUI. One issue I found is that the pieces move but don't "delete" if they are taken. This is a bug that I will hopefully fix by introducing a new function called upon once a piece is taken that deletes the piece from all instances. Ensuring it doesn't cause a crash where two pieces are tried to be moved at once.