GUI V0.22
Added a function to eliminate a piece from the list of its teams pieces. This now allows for pieces to be deleted from the game. However, there is an issue where they aren't getting deleted properly so I will need to revaluate the function, but I am close to completing it.