GUI - V0.3
Added a search for the secondary take, meaning once a piece is taken the search function is ran in case a second move is available. I did this by checking each potential move from the team who just took a piece and if A) a piece can be taken and B) the piece being able to take is in the same location as the piece that just moved, then second take available. I plan on creating the code for a second take now as It will be slightly different to the code for the usual take / move