Create GUI - V0.31
Created a function for the secondary move that is used once a second check is outputted as true. Added lots of comments to improve the clarity of my programme. Found a minor issue, I am unsure what causes it however I know it is linked to the secondary move and not switching players or eliminating a piece correctly