GUI - V033
Added a few clear functions in the secondary move function in order to erase "ghost" pieces from the screen, I believe the base game of the Checkers 2v2 is now functioning and I have tested with a friend to completion. Previous versions of the code crashed due to the secondary take error involving the function not looping until a move was made.