GUI - V-04
Implemented code to use the minimax algorithm to make AI move (compMove). Had to add some return calls to the Minimax code in order to pass back the move made, so I can update all the Boards regarding the GUI. I still need to create a new function that can eliminate pieces with the AI. I need to be careful how I implement it in case it breaks with the secondaryMove function.