Create GUI - V041
Created a procedure to eliminate a piece from the [team]_pieces list when a computer takes. Ran into one issue that had me confused but I realised when comparing the piece that had been taken and the current piece in the list both coordinates needed to be strings, fixing this and the procedure now works. A game can be completed against the AI, including with it having secondary takes to completion