GUI - V042
Using backtracking I decided to convert all of the pieces into objects, this is so I could use set an attribute isKing in order to check whether a piece is King or not, this ensures I can correctly display a piece as being a king. Ensuring the player knows which piece is king. However, there is an issue where the piece is only identified as King if it lands on the opposite square and stays there. If it conducts a secondary take and hits the square but doesn't stay, it will remain its original colour