Original code - V3
Fixed an issue where computer moves where not being displayed properly. In case of a secondary take, the program would only list the first jump as the whole move. Once fixed, the coordinates properly lined up in order for the kingCheck function in the GUI code to work, fixing the issue with the secondaryTakes and kings.