MM - V05
Added a play page which gives the user the choice of multiplayer, or single player with a variety of difficulties. I ran into some issues with calling on the GUI code which I changed by turning the whole thing into a function. Furthermore, I spotted some flaws with the AI algorithm so I will be fixing that soon. It was slow and inaccurate. With the main menu, it took awhile for me to get round all of the coordinates and locations of a 600 x 600 window however towards the end I had a good grasp of where I was placing the pieces. I had one major bug where specific boxes or texts weren't displaying properly but it came down to my if statements needing to be elifs and the coordinates of the objects.