Create MM - V11
Finished the settings page so each setting is available. I was accidentally adding and taking away 2 (when referring to the selected border) when I should have been doing the inverse as I forgot that a positive x value moves right and a positive y value moves down not the other way round.

Now the settings page has a save button, which returns the values and sets the appropriate colours. The colours are now passed as parameters into the GUI program so the GUI can let the pieces and board be the desired colours. I had to add new constant colours to stop any overlap due to variable names and ensure that text stayed the correct colour, etc. Settings page is now practically complete, I can still tweak RGB values and layout relatively easily so it is very adaptable code.