Create MM - V12
Added new colour pre-sets that represent each potential colourway, the boxes get highlighted in settings due to the highlight triangle function. The players choices are then fed into the CheckersGUI module where the colours are changed and have an affect on the board and piece colours.

At first, you could select the same colourway for piece and board but after testing I found it made the secondary player (white by default) invisible as the board and piece were the same exact colour. To fix this I removed the possibility to have the same board and piece colourway, only one.