Create LogIn - V0.1
First function, creates a list of all of the usernames string values, the second function compiles these lists into the individual usernames as list entries, then a linear search is conducted to match the given input to the potential usernames. That is all I have so far. Currently this only operates in the terminal. I plan on getting it all down in the terminal before converting it to work in the GUI. Next I will create functions to call upon the specific details of the users data.