Create LogIn - V02
Fixed the bubble sort algorithm after lots of testing, the sort was only completing one pass when in reality it needed multiple. This turned out to be because its terminating condition was wrong and needed tweaking so that it conducted multiple searches. I changed from boolean algebra to a count system, so each pass counted how many times something had been swapped. This meant it would now not continually switch the boolean to false, which triggered the terminating condition.

I have now started development on the file sort, which will reorder the file into alphabetical order (to enable a binary search). The current issue is that I only have one list, the ordered list with the correct line numbers. This means when the file is swapped with the correct line numbers, it swaps but is then overwritten when the temp value is reattributed. This happens because the first value in the sorted version will swap with "line 1" which is itself, meaning nothing changes. I hope to fix this by adding a secondary list.