Create LogIn - V03
Changed to an iterative sort model for the sort file function, this is a little slower but is useful as it means I am not wasting memory holding  temporary values. Now it functions well and you can search for each line once sorted. I am now hoping to implement searches for individual elements of data for the stats page