Create LogIn - V04
Fixed the iteration sort by making sure values that are in the correct position are still copied across (they previously were not). This now functions as wanted and sorts the file correctly. 

Started to develop the GUI log in page, it consists of a title and a main text box that I am currently working on. So far, it only accepts the letter a as I've been using that as a proof of concept, One issue is that you currently can't delete values and so the string grows until it goes off screen. Also I need to add the ability to enter the username and for it to then be added to the file.