Create LogIn - V05
Firstly, I changed the KEYDOWN == K_A selection statement as I would have to write a function for every single available value and it would take lots of time and be memory inefficient. Instead I used the event.unicode value that detects the unicode value that has just been input. I then disabled the unicode values $ and & from being entered as that would mess with my backend code. I need to remove the \ value so no one can enter a return value into the string, breaking the file. 

I then changed the Log in GUI code into a function that returns the username, this is because I can use recursion now to repeatedly check if the username matches the first entered username - ensuring the correct username is entered. 

Now, the functions return correctly and close and halt the GUI. This is because the plan is to now update the backend and then load the Main menu with the player logged in, stats ready to be updated.