Create LogIn - V06
Created the binary search algorithm that checks against the file. It uses the midpoints to discard parts of the list that the value cannot be found in. Currently returns true or false if it is found or not. I had many errors such as the empty username error (see screenshot) where I was accidentally making the username empty before every recursion, so the list was empty and crashed the software. Then I forgot to add brackets to the .isalpha function which meant I couldn't type in the text box as I wasn't checking if it was True or not. Now I need to code adding to the file if not found and the stats page if found and logged in (read specific line number)