Create Login - V07
Created code that if the username is found it will start the main menu (not implemented yet) and if the username is not found will append to the end of the file and then sort the file. (made the whole sorting portion of code into its own procedure (sortProcedure). However, the sort is not taking into account values that have the same starting alphabet value ( two J names) or how to handle if one contains capitals and one does not.