Create Login - V09
Firstly, I tested and added some black highlighting to the first page of the log in screen. I really like how it looks and am going to replicate it for all of the screens in the main menu. It is simply a hollow black rectangle on all of the box objects I am creating. 

Then I fixed an error with the greaterTHAN function where if the entered username was similar and shorter than a pre-existing username (for example, Aa compared with Aaron). This caused a list index out of range error as the compare wouldn't end before the end of the list was reached. Thus I made it so that if the function is still comparing by the time the end of the list is reached, say it is not greater than and place behind. This means aa would just go behind aaron and the binary search would still work.

Finally, I changed the lineNumber function so that it works with the new formatting of the data and now the line number of specific usernames can be gathered and sent to the main menu program.