Create MM - V13
I first made sure every element was highlighted with a black outline to make them all pop. I did this by creating a hollow black rectangle around the outside of each individual element. Then I started to implement the stats page. To do this I had to make the whole Main menu program into a procedure that is called by the log in program. The Login program passes two parameters, username and line number. Username is used for displaying the username on the stats page while line number is useful for accessing the data on the specified line for that user. I have simplified the data I will be monitoring into total wins, total losses,  win percentage,  and the result of the last 3 games. I think this is enough information to meet the brief of showing progress over time. Now, I only have a few features left to complete: I need to fix a bug where you can't make a typo in the log in stage without it crashing when returning values (recursive error), I need to make it so stats update when a game is completed. A second player can log in for multiplayer games and finally whoever exits the game is awarded a loss on their record.