Create GUI - V1
Latest version before final tests, I now pass User1 and User2 through the function so that the results of the games can be returned back to the main menu function. Removed the pygame.quit() when the game ends so that the user is instead taken back to the main menu rather than closing all software.