Create Login - V1
The main change was how I was calling the CheckersMM module. Now, I have made CheckersMM the main module and it calls upon the Login program. This change was necessary as there was many procedures I had to use in the CheckersMM program to log in but it caused a circular import error. Therefore Log in is now imported by the CheckersMM program. I then deleted all old and unused procedures from the old system of sort and searches I was using. I made it so the binary search only registered usernames as non-capital letters to ensure no (not found) searches occurred when one capital was changed. Therefore my log in is now not capital sensitive.

For the multiplayer game mode, a second log in occurs and passes a boolean value called secondLogin as True to the log in function, now it displays a unique text for the second log in to make it more clear to the user what is occurring. I removed all pygame.quit() from the function so the user is returned to the main menu after a log in. As previously if you closed pygame it would close it in the main menu function as well, crashing the program. Changed the binary search midpoint finder to more accurately represent the midpoint of the list in a 0 indexed format, as it used to be 1 indexed and caused many overflow errors. I then moved the not found check to the start of the binary search so it checks if the value is present first, rather then at the end where the list will have been changed further. Still need to complete more testing with extreme values but it currently functions to the quality required.