Create MM - V14
Made main 
Update data
Update line
Separate list
Reformat list

Firstly, I made the main menu program the central program, it imports the CheckersLogin and CheckersGUI functions in order to operate with all three. This is so the Main menu program can call on the Login function when multiplayer is chosen. Once I sorted the second Log in, I had to form functions to update the file with the results of the game. I create the update data function, which creates a list with all the file lines, and then edits the ones where the line number is equal to the index of the file list. To update the file line, the update line function is used. It uses another function to separate that specific file line into multiple values in a new list, based upon the commas in the line. Then, once all separated out the correct data is edited. For example, a win or loss and updating the last 3 games. Once the new file data has been changed in the  list. The final new function reformat list is used to make the separated back into a correct format and then re add it to the file.