GUI - V1.1
Found large error that was stopping the game to run correctly. After the users first move, the AI would not register the users first move and move independently of it. The users move also wouldn't appear on the screen but you could still click on where it should be and try and move it - causing an even larger error of the AI's logic. I found it was because the getmousemove function. I tested everything: copy() , deepcopy(), using multiple prints to understand what happened. None of it made sense. Inside the getmousemove function the board array was correct but outside after the return it suddenly changed and didn't match. Eventually, I tried calling on the CheckersAI.checkerBoard array that was being imported and should be up-to-date and correct. Somehow this managed to solve the issue even though the return was still incorrect. What's even stranger was that there was 0 issues with the multiplayer version of the code - which is no different to the singleplayer version. I am yet to understand the issue but this fix should suffice.